tests/wpt/web-platform-tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.object.invalidoffset.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.gradient.object.invalidoffset</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.gradient.object.invalidoffset</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var g = ctx.createLinearGradient(0, 0, 100, 0);
assert_throws_dom("INDEX_SIZE_ERR", function() { g.addColorStop(-1, '#000'); });
assert_throws_dom("INDEX_SIZE_ERR", function() { g.addColorStop(2, '#000'); });
assert_throws_js(TypeError, function() { g.addColorStop(Infinity, '#000'); });
assert_throws_js(TypeError, function() { g.addColorStop(-Infinity, '#000'); });
assert_throws_js(TypeError, function() { g.addColorStop(NaN, '#000'); });
t.done();

});
</script>
